+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## [Unreleased][unreleased]
 
 ### Added
+- added experimental tauri version (`packages/target-tauri`) #4462
 
 ### Changed
 - add some missing translations
 
 ### Fixed
+- fix cancelation of account deletion when canceling clicking outside of the dialog
 
 <a id="1_51_0"></a>